+++ packages/stream_chat/lib/src/core/models/event.dart
@@ -43,6 +43,7 @@ class Event {
     this.reminder,
     this.pushPreference,
     this.channelPushPreference,
+    this.channelMessageCount,
     this.extraData = const {},
     this.isLocal = true,
   }) : createdAt = createdAt?.toUtc() ?? DateTime.now().toUtc();
@@ -162,6 +163,9 @@ class Event {
   /// Push notification preferences for the current user for this channel.
   final ChannelPushPreference? channelPushPreference;
 
+  /// The total number of messages in the channel.
+  final int? channelMessageCount;
+
   /// Map of custom channel extraData
   final Map<String, Object?> extraData;
 
@@ -203,6 +207,7 @@ class Event {
     'reminder',
     'push_preference',
     'channel_push_preference',
+    'channel_message_count',
   ];
 
   /// Serialize to json
@@ -246,6 +251,7 @@ class Event {
     MessageReminder? reminder,
     PushPreference? pushPreference,
     ChannelPushPreference? channelPushPreference,
+    int? channelMessageCount,
     Map<String, Object?>? extraData,
   }) =>
       Event(
@@ -284,6 +290,7 @@ class Event {
         pushPreference: pushPreference ?? this.pushPreference,
         channelPushPreference:
             channelPushPreference ?? this.channelPushPreference,
+        channelMessageCount: channelMessageCount ?? this.channelMessageCount,
         isLocal: isLocal,
         extraData: extraData ?? this.extraData,
       );