+++ packages/stream_chat_persistence/lib/src/entity/channels.dart
@@ -39,6 +39,9 @@ class Channels extends Table {
   /// The count of this channel members
   IntColumn get memberCount => integer().withDefault(const Constant(0))();
 
+  /// The count of messages in this channel
+  IntColumn get messageCount => integer().nullable()();
+
   /// The id of the user that created this channel
   TextColumn get createdById => text().nullable()();